\name{DataMind-package}
\alias{DataMind-package}
\alias{DataMind}
\docType{package}
\title{Tools for the using the interactive www.DataMind.org learning platform}
\description{
Visit www.datamind.org for more information!
}
\details{
\tabular{ll}{
Package: \tab DataMind\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-18\cr
License: \tab What license is it under?\cr
}
}
\author{Jonathan Cornelissen <Jonathan@DataMind.org>}
\references{www.DataMind.org}
\keyword{ package }
